requirements-mac.txt
antlr4-python3-runtime==4.9.3
anyconfig==0.14.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dynaconf==3.2.10
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.35.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kedro==0.19.5
kedro-datasets==1.6.0
kedro-telemetry==0.6.2
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.1
notebook_shim==0.2.4
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip-tools==7.4.1
platformdirs==4.3.7
pluggy==1.2.0
pre_commit_hooks==5.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rope==1.13.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toposort==1.10
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.115
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
zipp==3.21.0
flask
kedro-datasets
easyocr